requirements.txt
# FILE: northstar/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
pydantic==2.8.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
alembic==1.13.2
redis==5.0.7
rq==1.16.2
httpx==0.27.2
requests==2.32.3
qdrant-client==1.15.1
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
pyjwt[crypto]==2.9.0
python-dotenv==1.0.1
loguru==0.7.2
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp==1.29.0
sentence-transformers==3.0.1
torch>=2.2.0
rapidfuzz==3.9.7
blake3>=0.4.1       
PyPDF2>=3.0.0 
protobuf>=5.28.3
grpcio>=1.66.2
googleapis-common-protos>=1.66.0
google-api-core>=2.20.0

google-auth>=2.35.0
pytest
pytest-asyncio
httpx
anyio
python-dotenv
pytest-cov
